More gde names refactored into people in xml layouts
