+++ app/src/main/java/org/gdg/frisbee/android/about/ContributorsFragment.java
@@ -16,42 +16,28 @@
 
 package org.gdg.frisbee.android.about;
 
-import android.content.Intent;
-import android.net.Uri;
 import android.os.Bundle;
 import android.support.design.widget.Snackbar;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ListView;
 
 import org.gdg.frisbee.android.Const;
 import org.gdg.frisbee.android.R;
 import org.gdg.frisbee.android.api.model.ContributorList;
-import org.gdg.frisbee.android.api.model.GdgPerson;
 import org.gdg.frisbee.android.app.App;
 import org.gdg.frisbee.android.cache.ModelCache;
-import org.gdg.frisbee.android.common.GdgListFragment;
 import org.gdg.frisbee.android.common.PeopleAdapter;
+import org.gdg.frisbee.android.common.PeopleListFragment;
 import org.gdg.frisbee.android.utils.Utils;
 import org.gdg.frisbee.android.view.ColoredSnackBar;
 import org.joda.time.DateTime;
 
-import butterknife.ButterKnife;
 import retrofit.Callback;
 import retrofit.RetrofitError;
 
-public class ContributorsFragment extends GdgListFragment {
-
-    protected PeopleAdapter mAdapter;
+public class ContributorsFragment extends PeopleListFragment {
 
     @Override
     public void onActivityCreated(Bundle savedInstanceState) {
-        super.onActivityCreated(savedInstanceState);
-
         mAdapter = new PeopleAdapter(getActivity());
-        setListAdapter(mAdapter);
-
         loadContributors();
     }
 
@@ -95,20 +81,5 @@ public class ContributorsFragment extends GdgListFragment {
 
             }
         });
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View v = inflater.inflate(R.layout.fragment_gde_list, container, false);
-        ButterKnife.bind(this, v);
-        return v;
-    }
-
-    @Override
-    public void onListItemClick(ListView l, View v, int position, long id) {
-        GdgPerson contributor = mAdapter.getItem(position);
-        Intent i = new Intent(Intent.ACTION_VIEW);
-        i.setData(Uri.parse(contributor.getUrl()));
-        startActivity(i);
     }
 }
