+++ app/src/main/java/org/gdg/frisbee/android/common/PlainLayoutFragment.java
@@ -1,4 +1,4 @@
-package org.gdg.frisbee.android.gde;
+package org.gdg.frisbee.android.common;
 
 import android.os.Bundle;
 import android.support.annotation.LayoutRes;
@@ -7,7 +7,6 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import org.gdg.frisbee.android.R;
-import org.gdg.frisbee.android.common.BaseFragment;
 
 /**
  * Created by maui on 03.06.2014.
