+++ app/src/main/res/layout/list_person_item.xml
@@ -24,8 +24,7 @@
         android:layout_height="wrap_content">
 
         <org.gdg.frisbee.android.widget.SquaredImageView
-          android:id="@+id/thumb"
-
+          android:id="@android:id/icon"
           android:layout_width="match_parent"
           android:layout_height="wrap_content" />
       </FrameLayout>
@@ -37,7 +36,7 @@
         android:padding="5dip">
 
         <TextView
-          android:id="@+id/name"
+          android:id="@android:id/text1"
           android:layout_width="match_parent"
           android:layout_height="wrap_content"
           android:ellipsize="end"
@@ -46,7 +45,7 @@
           android:textStyle="bold" />
 
         <TextView
-          android:id="@+id/country"
+          android:id="@android:id/text2"
           android:layout_width="match_parent"
           android:layout_height="wrap_content"
           android:ellipsize="end"
