+++ app/src/main/res/values-sw600dp-land/integers.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
   <integer name="news_fragment_column_count">2</integer>
-  <integer name="gde_fragment_column_count">4</integer>
+  <integer name="people_fragment_column_count">4</integer>
 </resources>
